src/requirements.txt
fastapi
fastapi-utils
requests
typing-inspect
python-dateutil
pytz
uvicorn
